refactor: Add Drawer
