+++ src/utils/addPlayerUtils.js
@@ -1,0 +1,13 @@
+export const getDefaultAddPlayerFilters = () => ({
+  team: '',
+  position: '',
+  offenseRole: '',
+  defenseRole: '',
+  subRoles: { offense: [], defense: [] },
+  shootingProfile: '',
+  badges: [],
+  minSalary: undefined,
+  maxSalary: undefined,
+  freeAgentYear: '',
+  freeAgentType: '',
+});
